Refactor protobufs.
For the first iteration we do not can focus on match and not
valiation. A unit is a prime number. Unity=1. We assume the
followig units and resources on the chain.
unity=1
bag=2
$=3
peanut=7
usd=9
iniate: 5 bag of peanut => [5][2/1][7/1]
reciprocate: 1$ of usd/1 bag of peanut =>
   [1/1][3/2][9/7]
match: 5 bag of peanut * 1$ of usd/1 bag of peanut = 5 $ of usd
   [5][2/1][7/1] * [1/1][3/2][9/7] = [5/1][3/1][9/1]

We'll deal with decimal precion and accuracy later.